refactor 'anchor_based' model loss
