+++ src/custom_yolo_lib/model/e2e/anchor_based/training_utils.py
@@ -2,8 +2,7 @@ import dataclasses
 from typing import Tuple
 import torch
 
-from custom_yolo_lib.model.e2e.anchor_based.constants import ANCHOR_GAIN
-import custom_yolo_lib.process.bbox.utils
+
 import custom_yolo_lib.model.building_blocks.heads.anchors.anchors_3_coco
 from custom_yolo_lib.model.building_blocks.heads.detections_3_anchors import (
     FeatureMapType,
@@ -50,167 +49,3 @@ def get_anchors_as_bbox_tensors(
         feature_map_type=FeatureMapType.LARGE,
     )
     return small_anchors, medium_anchors, large_anchors
-
-
-def build_feature_map_targets(
-    annotations: torch.Tensor,
-    anchor_tensor: AnchorsTensor,
-    grid_size_h: int,
-    grid_size_w: int,
-    num_classes: int,
-    check_values: bool = False,
-    epsilon: float = 1e-4,
-    is_ordinal_objectness: bool = False,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    anchors = anchor_tensor.anchors
-    num_anchors = anchors.shape[0]
-    assert num_anchors == 3, "Only 3 anchors are supported for each feature map type"
-
-    targets = torch.zeros(
-        (num_anchors, 5 + num_classes, grid_size_h, grid_size_w),
-        device=annotations.device,
-    )
-    targets_mask = torch.zeros(
-        (num_anchors, grid_size_h, grid_size_w),
-        dtype=torch.bool,
-        device=annotations.device,
-    )
-
-    if len(annotations) == 0:
-        return targets, targets_mask
-
-    # Extract annotation data
-    bx, by, bw, bh, class_ids = annotations.unbind(1)
-    class_ids = class_ids.long()
-
-    # Reshape for broadcasting
-    bw_expanded = bw.unsqueeze(1)  # [N, 1]
-    bh_expanded = bh.unsqueeze(1)  # [N, 1]
-    anchor_w = anchors[:, 2].unsqueeze(0)  # [1, num_anchors]
-    anchor_h = anchors[:, 3].unsqueeze(0)  # [1, num_anchors]
-
-    # Check which anchors are valid for each annotation
-    valid_mask = (
-        (anchor_w * (ANCHOR_GAIN**2) - epsilon >= bw_expanded)
-        & (anchor_w / (ANCHOR_GAIN**2) + epsilon <= bw_expanded)
-        & (anchor_h * (ANCHOR_GAIN**2) - epsilon >= bh_expanded)
-        & (anchor_h / (ANCHOR_GAIN**2) + epsilon <= bh_expanded)
-    )
-
-    # Get indices of valid pairs
-    ann_idx, anchor_idx = valid_mask.nonzero(as_tuple=True)
-
-    if len(ann_idx) == 0:
-        return targets, targets_mask
-
-    # Get corresponding data for valid pairs
-    sel_bx = bx[ann_idx]
-    sel_by = by[ann_idx]
-    sel_bw = bw[ann_idx]
-    sel_bh = bh[ann_idx]
-    sel_class_ids = class_ids[ann_idx]
-
-    # Calculate grid positions and offsets
-    grid_x = sel_bx * grid_size_w
-    grid_x_int = grid_x.int()
-    x_offset = grid_x - grid_x_int
-    grid_y = sel_by * grid_size_h
-    grid_y_int = grid_y.int()
-    y_offset = grid_y - grid_y_int
-
-    # Still need a loop for conditional update and _bump_objectness
-    objectness_score = 1.0
-    for i in range(len(ann_idx)):
-        a_idx = anchor_idx[i]
-        g_y = grid_y_int[i]
-        g_x = grid_x_int[i]
-
-        if targets[a_idx, 4, g_y, g_x] < objectness_score:
-            # Update bbox
-            targets[a_idx, 0, g_y, g_x] = x_offset[i]
-            targets[a_idx, 1, g_y, g_x] = y_offset[i]
-            targets[a_idx, 2, g_y, g_x] = sel_bw[i]
-            targets[a_idx, 3, g_y, g_x] = sel_bh[i]
-
-            if check_values:
-                assert 0 <= x_offset[i] <= 1
-                assert 0 <= y_offset[i] <= 1
-                assert 0 <= sel_bw[i] <= 1
-                assert 0 <= sel_bh[i] <= 1
-
-            # Update objectness
-            targets[a_idx, 4, g_y, g_x] = 1.0
-            if is_ordinal_objectness:
-                _bump_objectness(targets, a_idx, g_y, g_x, max_value=objectness_score)
-
-            # Update class
-            targets[a_idx, 5 + sel_class_ids[i], g_y, g_x] = 1.0
-
-            targets_mask[a_idx, g_y, g_x] = True
-    return targets, targets_mask
-
-
-def _bump_objectness(
-    targets: torch.Tensor,
-    potential_anchor: int,
-    grid_y: int,
-    grid_x: int,
-    max_value: float = 1.0,
-) -> None:
-    """
-    In-place: raises objectness scores around (grid_y, grid_x)
-    for the given anchor in `targets`.
-    """
-    device = targets.device
-    _, _, H, W = targets.shape
-
-    # 8×2 tensors of offsets for 1-pixel and 2-pixel neighbors
-    offsets1 = torch.tensor(
-        [[-1, -1], [-1, 0], [-1, 1], [0, -1], [0, 1], [1, -1], [1, 0], [1, 1]],
-        device=device,
-    )
-    offsets2 = torch.tensor(
-        [
-            [-2, -2],
-            [-2, -1],
-            [-2, 0],
-            [-2, 1],
-            [-2, 2],
-            [-1, -2],
-            [-1, 2],
-            [0, -2],
-            [0, 2],
-            [1, -2],
-            [1, 2],
-            [2, -2],
-            [2, -1],
-            [2, 0],
-            [2, 1],
-            [2, 2],
-        ],
-        device=device,
-    )
-
-    # helper to bump a whole neighborhood
-    def bump(offsets: torch.Tensor, thresh: float) -> None:
-        # compute absolute neighbor coords: (8,2)
-        base = torch.tensor([grid_y, grid_x], device=device)
-        neigh = offsets + base  # shape (8,2)
-        ys, xs = neigh[:, 0], neigh[:, 1]
-
-        # mask in‐bounds
-        valid = (ys >= 0) & (ys < H) & (xs >= 0) & (xs < W)
-        if not valid.any():
-            return
-
-        ys, xs = ys[valid], xs[valid]
-        # gather current objectness scores
-        curr = targets[potential_anchor, 4, ys, xs]
-        # find which need raising
-        to_raise = curr < thresh
-        if to_raise.any():
-            targets[potential_anchor, 4, ys[to_raise], xs[to_raise]] = thresh
-
-    # apply to 1-pixel neighbors (→0.75 * max_value) and 2-pixel neighbors (→0.50 * max_value)
-    bump(offsets1, thresh=max_value * 0.75)
-    bump(offsets2, thresh=max_value * 0.50)
